+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+#### 🚀 Updates
+
+- Updated `moon upgrade` to upgrade via proto if we detect that moon is managed by proto. This will
+  run `proto install moon latest`.
+
 #### 🐞 Fixes
 
 - Fixed the `moon upgrade` command not handling the new v2 distribution format correctly. If you are